SuperMarket-System
You can add or omit items, the price in "products.dat". Keep the "sales.dat" empty that will be your sales data in detail, also tell us how many products that was sold and total earnings. DO NOT CHANGE OR RENAME  THE FILES OR ELSE AN ERROR WILL BE OCCURED.